server/web/requirements.txt
fastapi
uvicorn[standard]
jinja2
PyJWT
httpx
python-multipart
